addedcommisioned photographer
