+++ metadata-editor/app/controllers/EditsApi.scala
@@ -45,7 +45,7 @@ object EditsApi extends Controller with ArgoHelpers {
     // to access the `val`s of the classes though without instantiating them.
     val usageRightsData =
       List(PrImage(), Handout(), Screengrab(), SocialMedia(), Obituary(), Pool(),
-           StaffPhotographer("?", "?"))
+           StaffPhotographer("?", "?"), CommissionedPhotographer("?", "?"))
         .map(CategoryResponse.fromUsageRights)
 
     respond(usageRightsData)
